Rename `qr.rs` to `multi_qr.rs`
