+++ rust/src/lib.rs
@@ -25,13 +25,13 @@ pub mod label_manager;
 mod manager;
 mod mnemonic;
 mod multi_format;
+mod multi_qr;
 mod network;
 mod node;
 mod node_connect;
 mod pending_wallet;
 mod psbt;
 mod push_tx;
-mod qr;
 mod redb;
 mod seed_qr;
 mod task;
