+++ rust/src/multi_qr.rs
